pythonコードファイルをsrc/からトップモジュールディレクトリへ移動
